requirements.txt
flask==3.0.0
supabase==2.3.4
python-dotenv==1.0.0
gunicorn==21.2.0
